[WIP] Re-implemented client API functionality in terms of byte buffers provided by the user.
This dissociates the implementation from sync or async HTTP responses.

Each API function now returns an `http::Request` for the user to execute.

The corresponding response type has a function to try to parse itself from a
byte slice provided by the user. If the function indicates that more data is
needed, the user must buffer up more data and call the function again. If the
function returns a parsed value, then the next step depends on the semantics of
the operation:

- If the operation is expected to return a single JSON value (get pods,
  list pods, etc), the user would either discard the buffer since there is
  nothing more to do.

- If the operation is expected to return multiple JSON values (watch pods, etc),
  the response will indicate how much of the buffer was consumed to parse a
  single value, and the user must retry with the remaining buffer to parse more
  values if they need them.

- If the operation returns a bare string (get pod logs), this works similar to
  the above case of multiple JSON values, except that the function returns
  string slices instead of JSON values. The returned offset should still be used
  to advance the buffer in preparation of parsing the next string slice.

Inspired by https://sans-io.readthedocs.io/how-to-sans-io.html

Fixes <TODO>